게시글 미디어 업로드 방식 수정

+++ src/main/java/com/becareful/becarefulserver/global/scheduler/S3CleanUpScheduler.java
@@ -1,5 +1,6 @@
 package com.becareful.becarefulserver.global.scheduler;
 
+import com.becareful.becarefulserver.global.properties.*;
 import com.becareful.becarefulserver.global.service.*;
 import lombok.*;
 import lombok.extern.slf4j.*;
@@ -11,9 +12,13 @@ import org.springframework.stereotype.*;
 @RequiredArgsConstructor
 public class S3CleanUpScheduler {
     private final S3Service s3Service;
+    private final S3Properties s3Properties;
 
     @Scheduled(cron = "0 0 0 * * *") // 메 00:00
     public void cleanupTempFiles() {
+        if (!s3Properties.getCleanupEnabled()) {
+            return;
+        }
         log.info("[S3CleanupScheduler] temp 폴더 정리 시작");
         s3Service.cleanAllTempFolders();
         log.info("[S3CleanupScheduler] temp 폴더 정리 완료");
@@ -21,6 +26,9 @@ public class S3CleanUpScheduler {
 
     @Scheduled(cron = "0 0 0 * * *") // 매 00:00
     public void cleanupPermanentFiles() {
+        if (!s3Properties.getCleanupEnabled()) {
+            return;
+        }
         log.info("[S3CleanupScheduler] permanent 폴더 정리 시작");
         s3Service.deleteOrphanPermanentFiles();
         log.info("[S3CleanupScheduler] permanent 폴더 정리 완료");
